parts_things/requirements.txt
python-dotenv
langgraph
langchain-core
datasets
tqdm
numpy
